requirements.txt
ucimlrepo
pandas
numpy
scikit-learn
sqlalchemy 
psycopg2-binary 
apache-airflow